Expected V and C conditional on 2019 (2007) JAEEMK #18
THe 2019 and 2007 problems share the same function, have the same idea, 
which is to solve for the expected value of V and C in 2008 and 2020 
from the perspective of 2007 and 2019, where the planner's observable 
information come from. 

There was an older slower function that vignette was based on. Updated 
to FOC based function for the 2019 problem, and added new 2007 tester.

Minimal changes to the SNW_EVUVW19_JMKY problem, which averages over 
JAEEMK states for JMKY states, which are observable to the planner. No 
separate 07 tester for the JMKY function, since 07 and 19 calls are 
identical.