Updating all checks function to allow for bush checks solution #18
bush checks steps:

1. solve for steady-state
2. solve for unemployed state in 2009 given xi and b for 2009
3. solve for 2008 v and policy given 2009 unemployment uncertainty
4. adjust the asset state based on value and consumption from (3) to see 
the effect of stimulus checks on 2008 v and c
5. take expectation over JAEEMK (age, savings, education, income shock, 
marital, kids) over 2008 conditional on 2007, and then integerate in 
2007 conditional on JMKY

Bush stimulus problem is faster than biden stimulus problem because the 
stimulus checks are evaluated in 2008 given V and C in 2008. For the 
Biden and Trump stimulus problems, stimulus is given ex-post of shock 
realization to both individuals with and without unemployment shocks, so 
have to evaluate stimulus effects for both employed and unmployed type 
households.